UHF-11186: Combined the plugin tests to a single test.
